Move logging code to separate file
